backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
motor==3.3.2
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
httpx==0.25.2
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
openai==1.3.7
python-dateutil==2.8.2
cors==1.0.1
fastapi-cors==0.0.6
ldap3==2.9.1
psutil==5.9.6
schedule==1.2.0
pywin32==306